Debug why mapdir doesn't work

+++ lib/cli/src/commands/create_exe.rs
@@ -2274,8 +2274,10 @@ mod http_fetch {
     }
 
     pub(super) fn untar(tarball: &Path, target: &Path) -> Result<Vec<PathBuf>> {
-        let _ = std::fs::remove_dir(target);
-        wasmer_registry::try_unpack_targz(tarball, target, false)?;
+        if !target.exists() {
+            let _ = std::fs::remove_dir(target);
+            wasmer_registry::try_unpack_targz(tarball, target, false)?;
+        }
         Ok(list_dir(target))
     }
 }
